+++ app/Http/Controllers/EpisodeController.php
@@ -11,6 +11,7 @@ use App\Http\Requests\UpdateEpisodeRequest;
 use App\Http\Requests\UpdateWhereEpisodeRequest;
 use App\Models\Season;
 use App\Models\Series;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class EpisodeController extends Controller
@@ -20,8 +21,12 @@ class EpisodeController extends Controller
      */
     public function index()
     {
-        $Episode = Episode::get();
-        return $this->response(code: 200, data: $Episode);
+        if (Auth::user()->block === 0) {
+            $Episode = Episode::get();
+            return $this->response(code: 200, data: $Episode);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
 
     /**
@@ -37,35 +42,39 @@ class EpisodeController extends Controller
      */
     public function store(StoreEpisodeRequest $request, StoreWhereSeason_idRequest $Season, StoreWhereSeriesRequest $Series)
     {
-        //validate
-        $request = $request->validated();
-        $Season = $Season->validated();
-        $Series = $Series->validated();
-        //Series_id
-        $Series_id = Series::all()->where('series_name', $Series['series_name'])->first()->id;
-        //seasons_id
-        $season_id = Season::where([
-            ['series_id', $Series_id],
-            ['season_number', $Season['season_number']],
-        ])->first()->id;
-        $request['season_id'] = $season_id;
-        //insert data
-        $insert = Episode::create([
-            'episode_number' => $request['episode_number'],
-            'description' => $request['description'],
-            'episode_url' => $request['episode_url'],
-            'season_id' => $request['season_id'],
-        ]);
-        //add num_of_seasons in table series
-        $num_of_episodes = Season::all()->where('id', $insert['season_id'])->first()->num_of_episodes;
-        $num_of_episodes = $num_of_episodes + 1;
-        $update_num_of_seasons = DB::table('seasons')->where('id', $insert['season_id'])->update([
-            'num_of_episodes' => $num_of_episodes
-        ]);
-        if ($insert) {
-            return $this->response(code: 201, data: $insert);
+        if (Auth::user()->block === 0) {
+            //validate
+            $request = $request->validated();
+            $Season = $Season->validated();
+            $Series = $Series->validated();
+            //Series_id
+            $Series_id = Series::all()->where('series_name', $Series['series_name'])->first()->id;
+            //seasons_id
+            $season_id = Season::where([
+                ['series_id', $Series_id],
+                ['season_number', $Season['season_number']],
+            ])->first()->id;
+            $request['season_id'] = $season_id;
+            //insert data
+            $insert = Episode::create([
+                'episode_number' => $request['episode_number'],
+                'description' => $request['description'],
+                'episode_url' => $request['episode_url'],
+                'season_id' => $request['season_id'],
+            ]);
+            //add num_of_seasons in table series
+            $num_of_episodes = Season::all()->where('id', $insert['season_id'])->first()->num_of_episodes;
+            $num_of_episodes = $num_of_episodes + 1;
+            $update_num_of_seasons = DB::table('seasons')->where('id', $insert['season_id'])->update([
+                'num_of_episodes' => $num_of_episodes
+            ]);
+            if ($insert) {
+                return $this->response(code: 201, data: $insert);
+            } else {
+                return $this->response(code: 400, data: 'Can\'t create ');
+            }
         } else {
-            return $this->response(code: 400, data: 'Can\'t create ');
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
         }
     }
 
@@ -74,13 +83,17 @@ class EpisodeController extends Controller
      */
     public function show(Episode $episode)
     {
-        if ($this->Subscripe()) {
-        $id = $episode->id;
-        $episode = Episode::with('seasons', 'episode_watchings')->find($id);
-        return $this->response(code: 200, data: $episode);
-    }else{
-        return 'you can\'t watch this episode untill pay it';
-    }
+        if (Auth::user()->block === 0) {
+            if ($this->Subscripe()) {
+                $id = $episode->id;
+                $episode = Episode::with('seasons', 'episode_watchings')->find($id);
+                return $this->response(code: 200, data: $episode);
+            } else {
+                return 'you can\'t watch this episode untill pay it';
+            }
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
 
     /**
@@ -94,36 +107,40 @@ class EpisodeController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateEpisodeRequest $request, Episode $episode, StoreWhereSeason_idRequest $Season, StoreWhereSeriesRequest $Series,UpdateWhereEpisodeRequest $episode_old)
+    public function update(UpdateEpisodeRequest $request, Episode $episode, StoreWhereSeason_idRequest $Season, StoreWhereSeriesRequest $Series, UpdateWhereEpisodeRequest $episode_old)
     {
-        //validate
-        $Season = $Season->validated();
-        $Series = $Series->validated();
-        $request = $request->validated();
-        $episode_old=$episode_old->validated();
-        //Series_id_new
-        $Series_id_new = Series::all()->where('series_name', $Series['series_name'])->first()->id;
-        //seasons_id_new
-        $season_id_new = Season::where([
-            ['series_id', $Series_id_new],
-            ['season_number', $Season['season_number']],
-        ])->first()->id;
-        $request['season_id'] = $season_id_new;
-        //seasons_id_old
-        //Series_id_old
-        $Series_id_old = Series::all()->where('series_name', $episode_old['series_name_old'])->first()->id;
-        //seasons_id_old
-        $season_id_old = Season::where([
-            ['series_id', $Series_id_old],
-            ['season_number', $episode_old['season_number_old']],
-        ])->first()->id;
-        //update data
-      $update= DB::table('episodes')->where([['episode_number',$episode_old['episode_number_old']],['season_id',$season_id_old]])->update($request);
-      if ($update) {
-        return $this->response(code: 201, data: $update);
-    } else {
-        return $this->response(code: 400, data: 'Can\'t update ');
-    }
+        if (Auth::user()->block === 0) {
+            //validate
+            $Season = $Season->validated();
+            $Series = $Series->validated();
+            $request = $request->validated();
+            $episode_old = $episode_old->validated();
+            //Series_id_new
+            $Series_id_new = Series::all()->where('series_name', $Series['series_name'])->first()->id;
+            //seasons_id_new
+            $season_id_new = Season::where([
+                ['series_id', $Series_id_new],
+                ['season_number', $Season['season_number']],
+            ])->first()->id;
+            $request['season_id'] = $season_id_new;
+            //seasons_id_old
+            //Series_id_old
+            $Series_id_old = Series::all()->where('series_name', $episode_old['series_name_old'])->first()->id;
+            //seasons_id_old
+            $season_id_old = Season::where([
+                ['series_id', $Series_id_old],
+                ['season_number', $episode_old['season_number_old']],
+            ])->first()->id;
+            //update data
+            $update = DB::table('episodes')->where([['episode_number', $episode_old['episode_number_old']], ['season_id', $season_id_old]])->update($request);
+            if ($update) {
+                return $this->response(code: 201, data: $update);
+            } else {
+                return $this->response(code: 400, data: 'Can\'t update ');
+            }
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
 
     /**
@@ -135,22 +152,38 @@ class EpisodeController extends Controller
     }
     public function delete(Episode $episode)
     {
-        $delete = $episode->delete();
-        return $this->response(code: 202, data: $delete);
+        if (Auth::user()->block === 0) {
+            $delete = $episode->delete();
+            return $this->response(code: 202, data: $delete);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
     public function deleted(Episode $episode)
     {
-        $deleted = $episode->onlyTrashed()->get();
-        return $this->response(code: 302, data: $deleted);
+        if (Auth::user()->block === 0) {
+            $deleted = $episode->onlyTrashed()->get();
+            return $this->response(code: 302, data: $deleted);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
     public function restore($episode)
     {
-        $episode = Episode::withTrashed()->where('id', $episode)->restore();
-        return $this->response(code: 202, data: $episode);
+        if (Auth::user()->block === 0) {
+            $episode = Episode::withTrashed()->where('id', $episode)->restore();
+            return $this->response(code: 202, data: $episode);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
     public function delete_from_trash($episode)
     {
-        $episode  = Episode::where('id', $episode)->forceDelete();
-        return $this->response(code: 202, data: $episode);
+        if (Auth::user()->block === 0) {
+            $episode  = Episode::where('id', $episode)->forceDelete();
+            return $this->response(code: 202, data: $episode);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
 }
