+++ app/Http/Controllers/EpisodeWatchingController.php
@@ -17,8 +17,12 @@ class EpisodeWatchingController extends Controller
      */
     public function index()
     {
-        $EpisodeWatching = EpisodeWatching::get();
-        return $this->response(code: 200, data: $EpisodeWatching);
+        if (Auth::user()->block === 0) {
+            $EpisodeWatching = EpisodeWatching::get();
+            return $this->response(code: 200, data: $EpisodeWatching);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
 
     /**
@@ -35,17 +39,21 @@ class EpisodeWatchingController extends Controller
      */
     public function store(StoreEpisodeWatchingRequest $request)
     {
-        if ($this->Subscripe()) {
-            //validate
-            $request=$request->validated();
-            //user_id
-            $user_id=Auth::user()->id;
-            $request['user_id']=$user_id;
-            //insert data
-            $insert=EpisodeWatching::create($request);
-            return $this->response(code:201,data:$insert);
-        }else{
-            return $this->response(code:500,msg:'not payment');
+        if (Auth::user()->block === 0) {
+            if ($this->Subscripe()) {
+                //validate
+                $request = $request->validated();
+                //user_id
+                $user_id = Auth::user()->id;
+                $request['user_id'] = $user_id;
+                //insert data
+                $insert = EpisodeWatching::create($request);
+                return $this->response(code: 201, data: $insert);
+            } else {
+                return $this->response(code: 500, msg: 'not payment');
+            }
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
         }
     }
 
@@ -54,9 +62,13 @@ class EpisodeWatchingController extends Controller
      */
     public function show(EpisodeWatching $episodeWatching)
     {
+        if (Auth::user()->block === 0) {
             $id = $episodeWatching->id;
             $episodeWatching = EpisodeWatching::with('users', 'episodes')->find($id);
             return $this->response(code: 200, data: $episodeWatching);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
 
     /**
@@ -84,22 +96,38 @@ class EpisodeWatchingController extends Controller
     }
     public function delete(EpisodeWatching $episodeWatching)
     {
-        $delete = $episodeWatching->delete();
-        return $this->response(code: 202, data: $delete);
+        if (Auth::user()->block === 0) {
+            $delete = $episodeWatching->delete();
+            return $this->response(code: 202, data: $delete);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
     public function deleted(EpisodeWatching $episodeWatching)
     {
-        $deleted = $episodeWatching->onlyTrashed()->get();
-        return $this->response(code: 302, data: $deleted);
+        if (Auth::user()->block === 0) {
+            $deleted = $episodeWatching->onlyTrashed()->get();
+            return $this->response(code: 302, data: $deleted);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
     public function restore($episodeWatching)
     {
-        $episodeWatching = EpisodeWatching::withTrashed()->where('id', $episodeWatching)->restore();
-        return $this->response(code: 202, data: $episodeWatching);
+        if (Auth::user()->block === 0) {
+            $episodeWatching = EpisodeWatching::withTrashed()->where('id', $episodeWatching)->restore();
+            return $this->response(code: 202, data: $episodeWatching);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
     public function delete_from_trash($episodeWatching)
     {
-        $episodeWatching  = EpisodeWatching::where('id',  $episodeWatching)->forceDelete();
-        return $this->response(code: 202, data: $episodeWatching);
+        if (Auth::user()->block === 0) {
+            $episodeWatching  = EpisodeWatching::where('id',  $episodeWatching)->forceDelete();
+            return $this->response(code: 202, data: $episodeWatching);
+        } else {
+            return $this->response(code: 401, msg: "You cannot log in because you are blocked.");
+        }
     }
 }
